Move loading the meta files to its own module and convert to TypeScript
